Implementation of okta login

+++ src/main/java/project_webtech/satu/repository/CoinwalletRepository.java
@@ -3,6 +3,6 @@ package project_webtech.satu.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import project_webtech.satu.entity.Coinwallet;
 
-public interface CoinwalletRepository extends JpaRepository <Coinwallet, Integer> {
+public interface CoinwalletRepository extends JpaRepository <Coinwallet, Long> {
 
 }
